+++ history.md
@@ -27,7 +27,8 @@
 - 添加createCustomEdit函数创建自定义组件，添加CustomLoadEdit组件
 - 全局添加DefaultEdit的泛型M
 - 优化createCustomEdit的返回值，修正类型报错
-- SearchData添加loaded/loadFail生命周期，对应字典构建完成和失败的生命周期
+- SearchData添加inited生命周期，对应字典构建完成生命周期
+- ComplexData添加$onSearchInited=>在检索加载完成后触发回调函数,loadDataBySearchInited=>在检索加载完成后触发数据加载函数
 
 ### 4.8.2/3
 - 添加TrackData轨迹数据结构